move expectation tests to unit tests folder
